requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
dj-database-url==2.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==23.2
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.3
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
